LayoutTests/fast/events/onunload-not-on-body-expected.txt

frame "<!--frame3-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 1: Use of window.alert is not allowed while unloading a page.
you should only see one unload alert appear.
